Frontend: Move Django and Apache to Singularity Image
Complete restructuring of the Carme-Frontend folder. Moved all original files to Carme-Django. Added new Folder Carme-Server-Conf which containces the hosts-file and all relevant apache-config files. Those are mounted in the the image and are simlinks to th respective files inside the image. Additionally is the folder containing the django python files renamed to scripts (originally called webfrontend).