SCRIPT: rewrote the entire script that creates the CarmeConfig_blanco
The script does not delete the entries of defined DEFAULT_VARIABLES and is sensitive to variables with double quotes, single quotes and without. Additionally an existing CarmeConfig_blanco file is deleted and a new one created with the same file rights as CarmeConfig. Note that if new variables are added to CarmeConfig we have to check if cleanCarmeConfig.sh has to be adjusted or not!