Job-Start: Sigularity Images are started inside a new pid namespace using newpid
With the help of this new pid namespace it is possible to jail every process running in a job within the batchjob. This makes it also for orphaned processes impossible to be adopted by init and therefore the cannot escape the batchjob anymore (^-^). Additinally, the use of newpid makes it impossible for the user to see in tools like htop the start syntax of e.g. theia, jupyter-lab or tensorboard. This has two advantages. First, this makes it easier for the users to identify their processes within htop. Second, from a security point of view we hide more perhaps security relevant things from normal users.