start proxy as www-data
the proxy is started as www-data now. In order to get this running we have to modify the traefik.toml to use not port 80 and 443 but ports larger that 1024 (binding of ports <1024 not allowed as normal user) AND redirect ports 80 and 443 to 10080 and 10443 on the LOGIN_NODE using iptablesgit add Carme-Scripts/infrastucture/run-carme-proxy.sh!