HOTFIX: removed additional empty line at the end of new toml files
for an unknown reason we have to remove the second empty line at the end of the job-toml files. I don't know why but if we have to many of these files the proxy breaks down. So far it seams to be a working fix, but we have to investigate this further.