GUI: Changed the color of the message visible on the webpage and the content of the messages.
The colors are modified in such a way that it is not misleading - Scheduled <-> Dark Yellow, Started <-> Green, Terminated <-> Blue. Additionally the messages include now date and time, the order of the messages in the message block is now latest first and the message block is scrolable for the last 10 messages.