FEATURE: Support for heterogeneous GPU nodes.
This is the initial commit needed in order to support heterogeneous systems. This includes many changes in contrast to what we had before. In order to schedule the GPUs we use now SLURMs implementation. The advantage is simply that we can use a lot of other features as well - e.g. defining the GPU type and the binding between GPU and CPU. As a consequence we have to modify essentially all scripts that are part of the 'Job-Start-Process' as well. By now the users can choose between 'I don't care which GPU I get' (default) and the specific GPU Type. This defined via a new environment variable CARME_GPU_TYPE = default, type-01, type-02, ... . Within the same process we modified the messages visible in the Webfrontend as well (different colors, time stamps, a scrollbar and inverting the displayed order). Not solved so far is the situation with the number of CPU cores and the reserved main memory. This remains an ongoing discussion.