FEATURE: Define default values for predefined GPU-Types.
Allocate the CPUs and Memory according to predefined values in the CarmeConfig file. The solution at the moment is using a variable called CARME_GPU_DEFAULTS. In this variable we set defaults according to type:cpu:mem and read it at the moment we submit a job. In order to read this variable we use a function (utilizing grep and regex) and source only this specific variable. Doing it this way we have more control about the variables that are available in the current shell.