FEATURE REQUEST: Partially solves issue #46.
Note that overwriting the displayed browser tab name is hard. And does only works partially. In order to have full control over this one would have to edit the source code of the respective application. Nevertheless we can provide a partially sollution. Partially, because it depends on various factors like browser version and so on.