BUGFIX: each multinode and -gpu job starts its on ssh sever with different key.
With the ability to have more than one multinode and -gpu job we have to change the structure of the ssh server start. In order to have control over the different keys in different jobs each key is labeled with the respective job-ID. To make this work we need an alias for ssh that binds to the respective key. These keys are now deleted via the slurm epiloge script.