BUG: ssh config cannot contain the identity file.
This is a bit problematic as we use identity files to connect via ssh which a different for each job. The problem is that the config cannot depend on the SLURM_JOB_ID which makes it useless to write the identity file to the config as it is possible to have more than one multi-gpu or multi-node job on a node (or involving the same node).