corrects slurm-epilog-script error.
The epilog-script fails for multi-node jobs as we miss the timing to delete the folders. Wants to say, the script is executed on all nodes of the jobs and it seams to happen that in the moment the folder is deleted on one node the other nodes does not know this as at the moment the other node checked if the file exists it still existed but not the moment it tries to delete it. In order to overcome that problem we check now if the node is the master node of the job and if this is true it deletes the files in the user home but the other nodes does not.