get number of nodes and number of allocated GPUs via scontrol
Note that this has to be done this complicated way as in the node prolog SLURM-JOB-NODELIST is not availabe and CUDA-VISIBLE-DEVICES does not give any usefull output. The last one may be realated the masking of the GPU id inside a job. This is only an assumtion as the slurm config does not tell anything regarding this problem.