adding maintenance mode state file
As we build the frontend singularity container in the CARME base folder we include any existing maintenance mode state file. This might be only true if there was a CARME installation running that was not utilizing the frontend service (as it was the case for us) but it is possible. This way we make sure that the moment we start or restart the frontend service, the frontend is accessible. Additionally, by adding this file to the repo gives admins the freedom to decide if they want to start CARME with maintenance mode on or off.